+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - 🐛(front) fix target blank links in chat #103
 - 🚑️(posthog) pass str instead of UUID for user PK #134
 - ⚡️(web-search) keep running when tool call fails #137
+- ✨(summarize): new summarize tool integration #78
 
 ### Removed